Batch class, Texture Batching, Textures tab, Assets class
Singletone method getInstance() in Stage class instead of public INSTANCE field,
ByteStream method readShortArray changed return value type from int[] to short[],
ProgressDialog removed

also some refactoring
TODO: fix 6th texture type

+++ src/com/vorono4ka/streams/ByteStream.java
@@ -119,10 +119,10 @@ public class ByteStream {
         return array;
     }
 
-    public int[] readShortArray(int count) {
-        int[] array = new int[count];
+    public short[] readShortArray(int count) {
+        short[] array = new short[count];
         for (int i = 0; i < array.length; i++) {
-            array[i] = this.readShort();
+            array[i] = (short) this.readShort();
         }
 
         return array;
